mqtt-testing/scripts/requirements.txt
# Python dependencies for Sonos MQTT Bridge
# Install with: pip install -r requirements.txt

paho-mqtt>=1.6.1
soco>=0.29.1
requests>=2.28.0
